add three-d dependency
