+++ Cargo.toml
@@ -1,9 +1,16 @@
 [package]
 name = "plasm"
 version = "0.1.0"
-authors = ["Francesco Magliocca <franciman12@gmail.com>"]
+authors = ["Francesco Magliocca <franciman12@gmail.com>",
+            "Francesco Pham <pham.francesco@gmail.com"]
 edition = "2018"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[lib]
+crate-type = ["cdylib"]
 
 [dependencies]
+three-d = "0.1.0"
+log = "0.4"
+wasm-bindgen = "0.2"
+console_error_panic_hook = "0.1"
+console_log = "0.1"
