+++ assets/shaders/color.frag
@@ -1,0 +1,8 @@
+in vec3 col;
+
+out vec4 fragmentColor;
+
+void main()
+{
+    fragmentColor = vec4(col, 1.0f);
+}
